Update tests suite and documentation
