--- tests/bootstrap.php
@@ -1,34 +1,0 @@
-<?php
-
-require __DIR__ . '/../vendor/autoload.php';
-
-PHPUnit_Framework_Error_Deprecated::$enabled = false;
-
-class TestCase extends PHPUnit_Framework_TestCase
-{
-    protected function expectCallableOnce()
-    {
-        $mock = $this->createCallableMock();
-
-        $mock
-            ->expects($this->once())
-            ->method('__invoke');
-
-        return $mock;
-    }
-
-    /**
-     * @link https://github.com/reactphp/react/blob/master/tests/React/Tests/Socket/TestCase.php (taken from reactphp/react)
-     */
-    protected function createCallableMock()
-    {
-        return $this->getMockBuilder('CallableStub')->getMock();
-    }
-}
-
-class CallableStub
-{
-    public function __invoke()
-    {
-    }
-}
